+++ .github/workflows/ci.yml
@@ -34,7 +34,7 @@ jobs:
         run: pipenv run python3 -m pytest --junitxml=test-results.xml
 
       - name: Validate Python code with flake8
-        run: pipenv run python3 -m flake8 --select BLK *.py
+        run: pipenv run python3 -m flake8 --select BLK **/*.py
 
       - name: Upload test results
         uses: actions/upload-artifact@v3
