+++ b2500/__init__.py
@@ -1,0 +1,1 @@
+from .b2500 import B2500
