+++ powermeter/base.py
@@ -1,0 +1,8 @@
+# Powermeter classes
+class Powermeter:
+
+    def wait_for_message(self, timeout=5):
+        pass
+
+    def get_powermeter_watts(self):
+        raise NotImplementedError()
